--- src/soil/controllers/soil-report.controller.ts
@@ -1,18 +1,0 @@
-import { Body, Controller, Get, Post } from '@nestjs/common';
-import { SoilReportService } from '../services/soil-report.service';
-import { CreateSoilReportDto } from '../dtos/create-soil-report-blockwise.dto';
-
-@Controller('soil-reports')
-export class SoilReportController {
-  constructor(private readonly service: SoilReportService) {}
-
-  @Post()
-  async create(@Body() dto: CreateSoilReportDto) {
-    return this.service.create(dto);
-  }
-
-  @Get()
-  async findAll() {
-    return this.service.findAll();
-  }
-}
