--- src/soil/modules/soil-report.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { SoilReport } from '../entities/soil-report-blockwise.entity';
-import { SoilReportService } from '../services/soil-report.service';
-import { SoilReportController } from '../controllers/soil-report.controller';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([SoilReport])],
-  providers: [SoilReportService],
-  controllers: [SoilReportController],
-})
-export class SoilReportModule {}
